Fixed formatting and PEP8 violations
Trying to get this project to work on a python2.7 system.
Ran into lots of tabs vs spaces issues, which I have cleaned up.
Also cleaned up a LOT of pep8 violations and python standard
coding practices to make the code readable and not complain.

It also seems to work now